RavenDB-17760 : address PR comments - move all sharded related functions into new partial classes e.g. AttachmentsStorage.Sharding, CountersStorage.Sharding, etc.
