+++ src/Raven.Server/Documents/AttachmentsStorage.cs
@@ -9,9 +9,7 @@ using Raven.Client.Exceptions;
 using Raven.Client.Exceptions.Documents;
 using Raven.Client.Exceptions.Documents.Indexes;
 using Raven.Server.Documents.Replication.ReplicationItems;
-using Raven.Server.Documents.Sharding;
 using Raven.Server.ServerWide.Context;
-using Raven.Server.Utils;
 using Sparrow;
 using Sparrow.Binary;
 using Sparrow.Json;
@@ -30,7 +28,7 @@ using Constants = Raven.Client.Constants;
 
 namespace Raven.Server.Documents
 {
-    public unsafe class AttachmentsStorage
+    public unsafe partial class AttachmentsStorage
     {
         internal static readonly TableSchema AttachmentsSchema = Schemas.Attachments.Current;
 
@@ -73,24 +71,6 @@ namespace Raven.Server.Documents
             foreach (var result in table.SeekForwardFrom(AttachmentsSchema.FixedSizeIndexes[AttachmentsEtagSlice], etag, 0))
             {
                 var attachment = TableValueToAttachment(context, ref result.Reader);
-
-                var stream = GetAttachmentStream(context, attachment.Base64Hash);
-                if (stream == null)
-                    ThrowMissingAttachment(attachment.Name);
-
-                attachment.Stream = stream;
-
-                yield return AttachmentReplicationItem.From(context, attachment);
-            }
-        }
-
-        public IEnumerable<ReplicationBatchItem> GetAttachmentsByBucketFrom(DocumentsOperationContext context, int bucket, long etag)
-        {
-            var table = context.Transaction.InnerTransaction.OpenTable(AttachmentsSchema, AttachmentsMetadataSlice);
-
-            foreach (var result in ShardedDocumentsStorage.GetItemsByBucket(context.Allocator, table, AttachmentsSchema.DynamicKeyIndexes[AttachmentsBucketAndEtagSlice], bucket, etag))
-            {
-                var attachment = TableValueToAttachment(context, ref result.Result.Reader);
 
                 var stream = GetAttachmentStream(context, attachment.Base64Hash);
                 if (stream == null)
@@ -1358,42 +1338,6 @@ namespace Raven.Server.Documents
                     yield return attachment;
                 }
             }
-        }
-
-        [StorageIndexEntryKeyGenerator]
-        internal static ByteStringContext.Scope GenerateBucketAndEtagIndexKeyForAttachments(ByteStringContext context, ref TableValueReader tvr, out Slice slice)
-        {
-            return ShardedDocumentsStorage.ExtractIdFromKeyAndGenerateBucketAndEtagIndexKey(context, (int)AttachmentsTable.LowerDocumentIdAndLowerNameAndTypeAndHashAndContentType,
-                (int)AttachmentsTable.Etag, ref tvr, out slice);
-        }
-
-        private void UpdateBucketStatsOnPutOrDeleteStream(DocumentsOperationContext context, Slice attachmentKey, long sizeChange)
-        {
-            if (_documentDatabase is not ShardedDocumentDatabase)
-                return;
-
-            using (GetBucketFromAttachmentKey(context, attachmentKey, out var bucketSlice))
-            {
-                ShardedDocumentsStorage.UpdateBucketStats(context.Transaction.InnerTransaction, bucketSlice, oldSize : 0, newSize : sizeChange);
-            }
-        }
-
-        private static ByteStringContext.Scope GetBucketFromAttachmentKey(DocumentsOperationContext context, Slice attachmentKey, out Slice bucketSlice)
-        {
-            int sizeOfDocId = 0;
-            for (; sizeOfDocId < attachmentKey.Size; sizeOfDocId++)
-            {
-                if (attachmentKey[sizeOfDocId] == SpecialChars.RecordSeparator)
-                    break;
-            }
-
-            var bucket = ShardHelper.GetBucket(attachmentKey.Content.Ptr, sizeOfDocId);
-            var scope = context.Allocator.Allocate(sizeof(int), out var buffer);
-            *(int*)buffer.Ptr = Bits.SwapBytes(bucket);
-
-            bucketSlice = new Slice(buffer);
-
-            return scope;
         }
     }
 }
